+++ src/main/scala/uk/co/gresearch/spark/dgraph/connector/TriplePartitionReader.scala
@@ -21,7 +21,7 @@ import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.connector.read.PartitionReader
 import uk.co.gresearch.spark.dgraph.connector.model.GraphTableModel
 
-class TriplePartitionReader(partition: Partition, model: GraphTableModel) extends PartitionReader[InternalRow] {
+case class TriplePartitionReader(partition: Partition, model: GraphTableModel) extends PartitionReader[InternalRow] {
 
   lazy val rows: Iterator[InternalRow] = model.modelPartition(partition)