+++ src/main/scala/uk/co/gresearch/spark/dgraph/connector/TripleTable.scala
@@ -22,15 +22,14 @@ import java.util.UUID
 import org.apache.spark.sql.connector.read.ScanBuilder
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.util.CaseInsensitiveStringMap
-import uk.co.gresearch.spark.dgraph.connector.encoder.TripleEncoder
 import uk.co.gresearch.spark.dgraph.connector.model.GraphTableModel
 import uk.co.gresearch.spark.dgraph.connector.partitioner.Partitioner
 
-class TripleTable(partitioner: Partitioner, model: GraphTableModel, val cid: UUID) extends TableBase {
+case class TripleTable(partitioner: Partitioner, model: GraphTableModel, val cid: UUID) extends TableBase {
 
   override def schema(): StructType = model.schema()
 
   override def newScanBuilder(options: CaseInsensitiveStringMap): ScanBuilder =
-    new TripleScanBuilder(partitioner, model)
+    TripleScanBuilder(partitioner, model)
 
 }