+++ src/main/scala/uk/co/gresearch/spark/dgraph/connector/partitioner/AlphaPartitioner.scala
@@ -27,7 +27,7 @@ case class AlphaPartitioner(schema: Schema, clusterState: ClusterState, partitio
 
   override def getPartitions: Seq[Partition] = {
     PredicatePartitioner.getPartitions(
-      schema, clusterState, getGroupTargets(clusterState, _).size * partitionsPerAlpha
+      schema, clusterState, getGroupTargets(clusterState, _).size * partitionsPerAlpha, Map.empty
     )
   }