+++ src/main/scala/uk/co/gresearch/spark/dgraph/connector/partitioner/SingletonPartitioner.scala
@@ -21,6 +21,6 @@ import uk.co.gresearch.spark.dgraph.connector.{Partition, Target}
 
 case class SingletonPartitioner(targets: Seq[Target]) extends Partitioner {
 
-  override def getPartitions: Seq[Partition] = Seq(Partition(targets, None, None))
+  override def getPartitions: Seq[Partition] = Seq(Partition(targets, None, None, None))
 
 }