+++ src/test/scala/uk/co/gresearch/spark/dgraph/connector/partitioner/TestAlphaPartitioner.scala
@@ -26,7 +26,7 @@ class TestAlphaPartitioner extends FunSpec {
 
   describe("AlphaPartitioner") {
 
-    val schema = Schema((1 to 7).map(i => Predicate(s"pred$i", s"type$i")).toSet)
+    val schema = Schema((1 to 7).map(i => Predicate(s"pred$i", s"type$i", s"type$i")).toSet)
     val clusterState = ClusterState(
       Map(
         "1" -> Set(Target("host1:9080")),
@@ -51,12 +51,12 @@ class TestAlphaPartitioner extends FunSpec {
       assert(partitions.length === 4)
       assert(partitions.toSet === Set(
         // predicates are shuffled within group and alpha, targets rotate within group, empty group does not get a partition
-        Partition(Seq(Target("host2:9080"), Target("host3:9080")), Some(Set(Predicate("pred1", "type1"), Predicate("pred2", "type2"))), None),
-        Partition(Seq(Target("host3:9080"), Target("host2:9080")), Some(Set(Predicate("pred3", "type3"), Predicate("pred4", "type4"))), None),
+        Partition(Seq(Target("host2:9080"), Target("host3:9080")), Some(Set(Predicate("pred1", "type1", "type1"), Predicate("pred2", "type2", "type2"))), None, None),
+        Partition(Seq(Target("host3:9080"), Target("host2:9080")), Some(Set(Predicate("pred3", "type3", "type3"), Predicate("pred4", "type4", "type4"))), None, None),
 
-        Partition(Seq(Target("host4:9080"), Target("host5:9080")), Some(Set(Predicate("pred5", "type5"))), None),
+        Partition(Seq(Target("host4:9080"), Target("host5:9080")), Some(Set(Predicate("pred5", "type5", "type5"))), None, None),
 
-        Partition(Seq(Target("host6:9080")), Some(Set(Predicate("pred7", "type7"), Predicate("pred6", "type6"))), None)
+        Partition(Seq(Target("host6:9080")), Some(Set(Predicate("pred7", "type7", "type7"), Predicate("pred6", "type6", "type6"))), None, None)
       ))
     }
 
@@ -68,15 +68,15 @@ class TestAlphaPartitioner extends FunSpec {
         assert(partitions.length === 7)
         assert(partitions.toSet === Set(
           // predicates are shuffled within group and alpha, targets rotate within group, empty group does not get a partition
-          Partition(Seq(Target("host2:9080"), Target("host3:9080")), Some(Set(Predicate("pred1", "type1"))), None),
-          Partition(Seq(Target("host2:9080"), Target("host3:9080")), Some(Set(Predicate("pred2", "type2"))), None),
-          Partition(Seq(Target("host3:9080"), Target("host2:9080")), Some(Set(Predicate("pred3", "type3"))), None),
-          Partition(Seq(Target("host3:9080"), Target("host2:9080")), Some(Set(Predicate("pred4", "type4"))), None),
+          Partition(Seq(Target("host2:9080"), Target("host3:9080")), Some(Set(Predicate("pred1", "type1", "type1"))), None, None),
+          Partition(Seq(Target("host2:9080"), Target("host3:9080")), Some(Set(Predicate("pred2", "type2", "type2"))), None, None),
+          Partition(Seq(Target("host3:9080"), Target("host2:9080")), Some(Set(Predicate("pred3", "type3", "type3"))), None, None),
+          Partition(Seq(Target("host3:9080"), Target("host2:9080")), Some(Set(Predicate("pred4", "type4", "type4"))), None, None),
 
-          Partition(Seq(Target("host4:9080"), Target("host5:9080")), Some(Set(Predicate("pred5", "type5"))), None),
+          Partition(Seq(Target("host4:9080"), Target("host5:9080")), Some(Set(Predicate("pred5", "type5", "type5"))), None, None),
 
-          Partition(Seq(Target("host6:9080")), Some(Set(Predicate("pred6", "type6"))), None),
-          Partition(Seq(Target("host6:9080")), Some(Set(Predicate("pred7", "type7"))), None)
+          Partition(Seq(Target("host6:9080")), Some(Set(Predicate("pred6", "type6", "type6"))), None, None),
+          Partition(Seq(Target("host6:9080")), Some(Set(Predicate("pred7", "type7", "type7"))), None, None)
         ))
       }
     )