+++ src/test/scala/uk/co/gresearch/spark/dgraph/connector/partitioner/TestGroupPartitioner.scala
@@ -26,7 +26,7 @@ class TestGroupPartitioner extends FunSpec {
 
   describe("GroupPartitioner") {
 
-    val schema = Schema((1 to 4).map(i => Predicate(s"pred$i", s"type$i")).toSet)
+    val schema = Schema((1 to 4).map(i => Predicate(s"pred$i", s"type$i", s"type$i")).toSet)
     val clusterState = ClusterState(
       Map(
         "1" -> Set(Target("host1:9080")),
@@ -48,8 +48,8 @@ class TestGroupPartitioner extends FunSpec {
 
       assert(partitions.length === 2)
       assert(partitions.toSet === Set(
-        Partition(Seq(Target("host2:9080"), Target("host3:9080")), Some(Set(Predicate("pred1", "type1"), Predicate("pred2", "type2"), Predicate("pred3", "type3"))), None),
-        Partition(Seq(Target("host4:9080")), Some(Set(Predicate("pred4", "type4"))), None)
+        Partition(Seq(Target("host2:9080"), Target("host3:9080")), Some(Set(Predicate("pred1", "type1", "type1"), Predicate("pred2", "type2", "type2"), Predicate("pred3", "type3", "type3"))), None, None),
+        Partition(Seq(Target("host4:9080")), Some(Set(Predicate("pred4", "type4", "type4"))), None, None)
       ))
     }