+++ src/test/scala/uk/co/gresearch/spark/dgraph/connector/partitioner/TestSingletonPartitioner.scala
@@ -31,7 +31,7 @@ class TestSingletonPartitioner extends FunSpec {
       val partitions = partitioner.getPartitions
 
       assert(partitions.length === 1)
-      assert(partitions.toSet === Set(Partition(targets, None, None)))
+      assert(partitions.toSet === Set(Partition(targets, None, None, None)))
     }
 
   }