+++ src/test/scala/uk/co/gresearch/spark/dgraph/connector/partitioner/TestUidRangePartitioner.scala
@@ -19,16 +19,16 @@ package uk.co.gresearch.spark.dgraph.connector.partitioner
 
 import java.util.UUID
 
+import javassist.bytecode.SignatureAttribute.ObjectType
 import org.scalatest.FunSpec
 import uk.co.gresearch.spark.dgraph.connector
-import uk.co.gresearch.spark.dgraph.connector.executor.JsonGraphQlExecutor
-import uk.co.gresearch.spark.dgraph.connector.{ClusterState, Json, Partition, Predicate, Schema, Target, Uid, UidRange}
+import uk.co.gresearch.spark.dgraph.connector.{ClusterState, Filters, ObjectTypeIsIn, ObjectValueIsIn, Partition, Predicate, PredicateNameIsIn, Schema, SubjectIsIn, Target, Uid, UidRange}
 
 class TestUidRangePartitioner extends FunSpec {
 
   describe("UidRangePartitioner") {
 
-    val schema = Schema(Set(Predicate(s"pred1", s"type1"), Predicate(s"pred2", s"type2")))
+    val schema = Schema(Set(Predicate("pred1", "type1", "type1"), Predicate("pred2", "type2", "type2")))
     val clusterState = ClusterState(
       Map(
         "1" -> Set(Target("host1:9080"), Target("host2:9080")),
@@ -69,7 +69,7 @@ class TestUidRangePartitioner extends FunSpec {
           assert(uidPartitions.length === partitions.length * ranges.length)
           val expectedPartitions = partitions.flatMap( partition =>
             ranges.zipWithIndex.map { case (range, idx) =>
-              Partition(partition.targets.rotateLeft(idx), partition.predicates, Some(range))
+              Partition(partition.targets.rotateLeft(idx), partition.predicates, Some(range), None)
             }
           )
 
@@ -106,6 +106,34 @@ class TestUidRangePartitioner extends FunSpec {
       assertThrows[IllegalArgumentException] {
         UidRangePartitioner(uidPartitioner, 1, UidCardinalityEstimator.forMaxLeaseId(1000))
       }
+    }
+
+    it("should support what decorated partitioner supports") {
+      val partitioner = PredicatePartitioner(schema, clusterState, 1)
+      val uidPartitioner = UidRangePartitioner(partitioner, 2, UidCardinalityEstimator.forMaxLeaseId(1000))
+
+      Seq(
+        Seq(SubjectIsIn(Uid("0x1"))),
+        Seq(PredicateNameIsIn("pred")),
+        Seq(PredicateNameIsIn("pred"), ObjectValueIsIn("value")),
+        Seq(ObjectTypeIsIn("type")),
+        Seq(ObjectValueIsIn("type"))
+      ).foreach {
+        filters =>
+          val actual = uidPartitioner.supportsFilters(filters)
+          val expected = partitioner.supportsFilters(filters)
+          assert(actual === expected, filters)
+      }
+    }
+
+    it("should forward filters to decorated partitioner") {
+      val partitioner = PredicatePartitioner(schema, clusterState, 1)
+      val uidPartitioner = UidRangePartitioner(partitioner, 2, UidCardinalityEstimator.forMaxLeaseId(1000))
+      val filters = Filters(Seq.empty, Seq.empty)
+      val actual =
+        uidPartitioner.withFilters(filters).asInstanceOf[UidRangePartitioner]
+          .partitioner.asInstanceOf[connector.partitioner.PredicatePartitioner].filters
+      assert(actual eq filters)
     }
 
   }