+++ CHANGELOG.md
@@ -6,10 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased] - YYYY-MM-DD
 
 ### Added
+- Added Spark filter pushdown to improve efficiency when loading only a subgraph. Filters like
+  `.where($"predicate" === "name")` will be pushed to Dgraph and only the relevant graph data will
+  be read ([issue #7](https://github.com/G-Research/spark-dgraph-connector/issues/7)).
 - Improve performance of `PredicatePartitioner` for multiple predicates per partition. Restoring
   default number of predicates per partition of `1000` from before 0.3.0 ([issue #22](https://github.com/G-Research/spark-dgraph-connector/issues/22)).
 - The `PredicatePartitioner` combined with `UidRangePartitioner` is the default partitioner now.
-- Add stream-like reading of partitions from Dgraph. Partitions are slip into smaller chunks.
+- Add stream-like reading of partitions from Dgraph. Partitions are split into smaller chunks.
   Works with predicate partitioning only, so it cannot be combined with uid partitioning.
 
 ### Security